Restore task dependency to standard 'build'
The geoclient-native post-build 'syncLocalNativeLibs' task was recently
changed to only sync native binaries to the common $localLibsDir for
native tasks which were 'buildable' on the current platform.

Although there's evidence that this causes specific problems, it seems
to be non-standard. Since this was changed to debug other issues since
solved, it seems prudent to restore more standard behavior until I
understand the Gradle C/C++ better.