Fix platform-specific headers in PrebuiltLibrary
Workaround to deal with an issue caused by the need for
platform-specific header files in a PrebuiltLibrary.
The geoclient-jni binaries use JNI and need to include
the platform-specific 'jni_md.h' file (found in the JDK
home under 'include/win32' or 'include/linux') in
addition to the standard 'jni.h' header under 'include'.

At the moment, it is affecting Gradle v2.14,
JDK 1.8.0_92-x86_64, MSYS2/mingw64 on Windows 7, but
it may be relevant for other platforms as well.

This fix works for guessable yet unknown reasons and
therefore should be considered a hack.

See: https://discuss.gradle.org/t/native-prebuilt-library-platform-specific-headers/17418