Working Docker build
Build which works with this project's Dockerfiles that rely on
mlipper/geosupport-docker Docker repository to provision a "data volume
container" libraries, datafiles and system library path configuration to
insure that Geosupport is installed in a known location.

Note, this commit will use gsIncludePath to find Geosupport's include
files if given, but defaults to a patched version of these headers
included in the geoclient-native/src/geosupport/headers directory.