[BROKEN] Initial upgrade to Spring 5.1.x
Almost all compile issues are resolved and almost all tests are passing. The
remaining issue is with geoclient-service project's use of JSONP
support.

Prior to Spring MVC 5.0.7, JSONP was supported by class
AbstractJsonpResponseBodyAdvice. Support for JSONP has been removed
completely in Spring MVC 5.1.0 and the recommended replacement is to use
the framework's CORS support.

This has the potential to break clients and must be documented and
announced appropriately.