[RESOLVED] Integration tests run w/Gradle passing
Issue turned out to be how Gradle tests were configured. The GEOFILES
variable was being used to set a Gradle property which was then returned
from a script method as a File.

When the File was converted back to a String it did NOT include the trailing
file separator. Thus, the trailing slash was missing when used to set the
GEOFILES variable within the "test { ... }" block.

This explains why the only test that passed was for function HR which
apparently doesn't use GEOFILES at all.

Coming up next...Move all the Groovy methods to a build plugin.