Fix gradle deprecation warning
After upgrading to Gradle 6.0.1 build gives a deprecation warning about
the use of 'baseName' instead of 'archiveBaseName' for Zip, Jar related
tasks.

This fixes one of the warnings in the src/doc/examples/jsonp example.
Another still exists and looks to be caused by the Spring Gradle plugin
when running the geoclient-service tasks. The Spring Gradle plugin
doesn't officially support Gradle 6.x yet so hopefully this will
disappear when it does.