Remove LatLongEnhancer, core Documentation APIs
Remove code and libraries for the following geoclient-service APIs:
1. LatLongEnhancer
This API is no longer necessary now that all Geosupport functions return
lat/long coordinates for those that used to be calculated using
LatLongEnhancer API. In addition to the API classes in
geoclient-service/src/main, the use of gov.nyc.doitt.gis.geometry
packages in geoclient-service/src/external and the dependence on
GeoTools libraries is also removed.

Note that because some clients may be relying on converted lat/long
coordinates which were returned as double values rather than strings,
the new FieldSetConverter API was introduced to continue this
conversion. The functions in question are 1B, AP, BL, BN, and 2W. For
fields, see the AppConfig class.

2. Use of the Documentation API in project geoclient-core
All use of the geoclient-core Documentation API (mostly classes in the
gov.nyc.doitt.gis.geoclient.doc package) has been removed from
geoclient-service.

The DocumentationController now relies only the Version class which
places it in the /doc view for use by Thymeleaf.