Working geoclient-service APIs post-spring upgrade.
This commit restores the full functionality of the existing Geoclient v2
REST and documentation web APIs following the upgrade of all Spring
Framework and Spring Boot dependencies. This includes all unit and
integration tests passing.

Note, however, that use of file extensions as part of endpoint URIs is
deprecated by Spring and will likely be removed from the framework in
upcoming releases. This is for good reason as there are legimate
security issues associated with this functionality.

Future releases of Geoclient will also deprecate and eventually remove this
behavior as well.