Move SFSController unit test to integrationTest
In the geoclient-service subproject, move and rename
SingleFieldSearchControllerTest.java from the src/test/java folder to
SingleFieldSearchControllerMvcIntegrationTest.java in the
src/integrationTest/java folder. Package remains the same.

Note: this test should ultimately be rewritten to use the correct
Spring/Spring Boot test API which properly isolates the web tier, this
is a temporary solution to align with upcoming use of Docker for
integration testing.